% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.p2c2m.snapp.R
\name{run.p2c2m.snapp}
\alias{run.p2c2m.snapp}
\title{Posterior predictive checks of SNAPP analyses}
\usage{
run.p2c2m.snapp(path_to_xml, xml_file = "snapp.xml", sum_stat = c("RF",
  "MLSD", "PFST"), num_sims = 100, dist_reps = 100,
  sample_unif = TRUE, delete_sims = FALSE, save_graphs = TRUE,
  save_output = TRUE, run_mode = 1)
}
\arguments{
\item{path_to_xml}{String. The path to the xml input files. Also the path that files will be written to}

\item{xml_file}{String. The .xml file in working directory. The empirical .xml file used as input for SNAPP.}

\item{sum_stat}{A string or vector of strings. Which summary statistics to test. RF = Robinson Foulds distance, MLSD = standard deviation of tree
likelihoods, PFST = Fst outlier test.}

\item{num_sims}{Numeric. Number of posterior predictive datasets to simulate.}

\item{dist_reps}{Numeric. Number of trees from the posterior predictive distribution to compare to the posterior tree the data was simulated from.}

\item{sample_unif}{Logical. If TRUE, sample posterior at regular interval. Otherwise, randomly sample.}

\item{delete_sims}{Logical. Should intermediate simulation files be deleted?}

\item{save_graphs}{Logical. Should graphs of each summary statistic be saved to files?}

\item{save_output}{Logical. Should output be saved to a file?}

\item{run_mode}{Numeric. Whether to simulate data and prepare xml files from the empirical data (1) OR calculate summary statistics and compare between    #'  empirical and simulated data (2)}
}
\value{
Posterior predictive datasets and .xml files will be created in the current working directory if run = 1. If run = 2, PDF files will be
 created showing histograms of each summary statistic, and two files will be created with a summary of the results.
}
\description{
\code{run.p2c2m.snapp} simulates posterior predictive datasets and compares them to the empirical dataset to identify model violations
}
\examples{
\donttest{
run.p2c2m.snapp(path_to_xml = tempdir(), xml_file = "snapp.xml",
                sum_stat = c("RF", "MLSD", "PFST"), num_sims = 100, dist_reps = 100,
                sample_unif = TRUE, delete_sims = FALSE, save_graphs = TRUE,
                save_output = TRUE, run_mode = 1)
run.p2c2m.snapp(path_to_xml = tempdir(), xml_file = "snapp.xml",
                sum_stat = c("RF", "MLSD", "PFST"), num_sims = 100, dist_reps = 100,
                sample_unif = TRUE, delete_sims = FALSE, save_graphs = TRUE,
                save_output = TRUE, run_mode = 2)
}

}
